Flytter ansvaret for umami-tracking til dekoratøren
