+++ package.json
@@ -26,7 +26,7 @@
         "@navikt/ds-css": "7.9.1",
         "@navikt/ds-react": "7.9.1",
         "@navikt/ds-tokens": "7.9.1",
-        "@navikt/nav-dekoratoren-moduler": "3.1.3",
+        "@navikt/nav-dekoratoren-moduler": "3.1.4-beta.1",
         "dayjs": "1.11.13",
         "html-react-parser": "5.2.2",
         "js-cookie": "3.0.5",
@@ -44,7 +44,6 @@
         "@types/node": "22.10.5",
         "@types/react": "18.3.12",
         "@types/react-dom": "18.3.1",
-        "@types/umami": "2.10.0",
         "@typescript-eslint/eslint-plugin": "8.19.1",
         "@typescript-eslint/parser": "8.19.1",
         "eslint": "8.57.0",
