+++ src/pages/_document.tsx
@@ -70,19 +70,13 @@ class MyDocument extends Document<Props> {
                     <meta name="twitter:description" content={description}/>
                     <meta name="twitter:image:src" content={previewImg}/>
                     <meta name="robots" content="noindex, nofollow"/>
-                    <script defer
-                            src="https://cdn.nav.no/team-researchops/sporing/sporing.js"
-                            data-host-url="https://umami.nav.no"
-                            data-website-id="c44a6db3-c974-4316-b433-214f87e80b4d"
-                    >
-                    </script>
-                    <Decorator.HeadAssets/>
+                    <Decorator.HeadAssets />
                 </Head>
                 <body>
-                    <Decorator.Header/>
-                    <Main/>
-                    <Decorator.Footer/>
-                    <Decorator.Scripts/>
+                    <Decorator.Header />
+                    <Main />
+                    <Decorator.Footer />
+                    <Decorator.Scripts />
                     <NextScript />
                 </body>
             </Html>
