+++ tsconfig.json
@@ -2,10 +2,7 @@
     "compilerOptions": {
         "target": "esnext",
         "lib": ["dom", "dom.iterable", "esnext"],
-        "types": [
-            "@testing-library/jest-dom",
-            "@types/umami"
-        ],
+        "types": ["@testing-library/jest-dom"],
         "allowJs": false,
         "skipLibCheck": true,
         "strict": true,
